Fixed and good cross validation

+++ submissions/starting_kit/feature_extractor.py
@@ -1,6 +1,6 @@
 from sklearn.base import BaseEstimator
 from sklearn.base import TransformerMixin
-
+import numpy as np
 
 class FeatureExtractor(BaseEstimator, TransformerMixin):
     def fit(self, X_df, y):
@@ -8,4 +8,4 @@ class FeatureExtractor(BaseEstimator, TransformerMixin):
 
     def transform(self, X_df):
         # get only the continuous variable
-        return X_df[["DEWP", "TEMP", "PRES"]].fillna(0)
+        return X_df[["DEWP", "TEMP", "PRES"]].astype(np.float).fillna(0)
